python/requirements.txt
appscript==1.3.0
certifi==2025.4.26
charset-normalizer==3.4.1
feedparser==6.0.11
idna==3.10
lxml==5.4.0
pillow==11.2.1
pyobjc-core==11.0
pyobjc-framework-Cocoa==11.0
pyobjc-framework-Quartz==11.0
python-dotenv==1.1.0
requests==2.32.4
sgmllib3k==1.0.0
urllib3==2.4.0
